wispr_bot/requirements.txt
aiogram
python-dotenv
asyncpg
pydantic
openai
pytest
pytest-asyncio
aiohttp
python-telegram-bot
alembic
loguru
